Fix bug of block marker setup script not being loaded; block recognition seems to be working correctly
The bug stems from a concurrency issue:
because we use `TemporaryDirectory` and `fork` at the same time,
the parent process would leave the `TemporaryDirectory` context and
remove the temporary zshrc config directory before the child process
could load the zshrc configuration.

After the fix, we're postponing the temporary directory removal by 10
seconds, as there is no (easy) way of knowing whether the child process has
finished setup in the parent process.
That could be brittle, especially if the child process takes more than
10 seconds to load the zshrc configuration,
but it's good enough for now.